Consolidate test, it's not even proper unit tests
